try pulling apart the validators a little and adding a type checking import clause to fix circular import
