+++ src/sentry/issues/grouptype.py
@@ -18,6 +18,7 @@ from sentry.features.base import OrganizationFeature
 from sentry.ratelimits.sliding_windows import Quota
 from sentry.types.group import PriorityLevel
 from sentry.utils import metrics
+from sentry.workflow_engine.endpoints.validators.error_detector import ErrorDetectorValidator
 
 if TYPE_CHECKING:
     from sentry.models.organization import Organization
@@ -261,7 +262,7 @@ class ErrorGroupType(GroupType):
     category = GroupCategory.ERROR.value
     default_priority = PriorityLevel.MEDIUM
     released = True
-    # TODO(cathy): add detector validator
+    detector_validator = ErrorDetectorValidator
     detector_config_schema = {"type": "object", "additionalProperties": False}  # empty schema
